Break out AbstractApp
`AbstractApp` is now the lifecycle Object with running event registery
`App` mixes in an "App Manager" to add child `App`s
