+++ docs/abstract-app.md
@@ -1,0 +1,30 @@
+# Marionette.Toolkit.AbstractApp
+
+`Marionette.Toolkit.AbstractApp` is an extention of `StateClass`.  Its purpose is to provide an object with a `initialize`/`start`/`stop`/`destroy` lifecycle.  All events bound to the `AbstractApp` while running (and only those) will be removed when stopped.
+
+**Note:** `Marionette.Toolkit.AbstractApp` class is not intended to be used directly. It will eventually become the `AbstractApp` class once the "App Manager" functionality is mixed in, instead of extending from the `AbstractApp`.
+
+## Documentation Index
+* [Lifecycle Methods](#lifecycle-methods)
+  * [start](#start)
+  * [stop](#stop)
+
+## Lifecycle Methods
+
+This
+
+### start
+
+This method...
+
+```js
+
+```
+
+### stop
+
+This method...
+
+```js
+
+```
