+++ docs/app. md
@@ -1,28 +1,3 @@
 # Marionette.Toolkit.App
 
-'Marionette.Toolkit.App' is
-
-## Documentation Index
-* [Lifecycle Methods](#lifecycle-methods)
-  * [start](#start)
-  * [stop](#stop)
-
-## Lifecycle Methods
-
-This
-
-### start
-
-This method...
-
-```js
-
-```
-
-### stop
-
-This method...
-
-```js
-
-```
+`Marionette.Toolkit.App` is an extention of `StateClass`.  Its purpose is to provide an object with a `initialize`/`start`/`stop`/`destroy` lifecycle.  All events bound to the `AbstractApp` while running (and only those) will be removed when stopped.
