+++ docs/state-class.md
@@ -1,5 +1,5 @@
 # Marionette.Toolkit.StateClass
 
-'Marionette.Toolkit.StateClass' is
+'Marionette.Toolkit.StateClass' is a `Marionette.Object` with a `Backbone.Model` for keeping state.
 
 ## Documentation Index
