+++ test/unit/abstract-app.spec.js
